fix: constructor type
