+++ src/vwbl/types/BiconomyConfigType.ts
@@ -1,0 +1,4 @@
+export type BiconomyConfig = {
+  apiKey: string;
+  forwarderAddress: string;
+};
