+++ src/vwbl/types/index.ts
@@ -4,3 +4,5 @@ export * from "./UploadMetadataType";
 export * from "./File";
 export * from "./EncryptLogic";
 export * from "./ProgressSubscriber";
+export * from "./BiconomyConfigType";
+export * from "./ConstructorPropsType";
